backend/requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.34.0
anthropic==0.49.0
gitpython==3.1.43
sqlalchemy==2.0.40
aiosqlite==0.20.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pydantic==2.11.0
httpx==0.28.1
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
stripe==11.3.0
cryptography==44.0.0
redis==5.2.1
fastembed==0.6.1
pgvector==0.3.6
